Refactors element initialization to use object
Updates element initialization to receive the object instead of the value.
This allows elements to dynamically access values through the object's
methods, providing more flexibility and enabling features like conditional
rendering based on the object's state.

Removes the need to pass the value directly during element instantiation.
Adds a `value` method to the Element class that retrieves the value from the object.

Removes primary key handling from the build_many_forms method in base and adds a primary_key element to the subform.
This element is only rendered when the object is persisted.