Adds default ID element to subforms
Ensures that subforms automatically include an ID element if one is not already defined.

This change simplifies subform creation by providing a default ID field, removing the need to manually define it in many cases.
It also removes now obsolete code that was used to add an abstract primary key element.