Adds support for detached elements in subforms
Allows subforms to handle elements that are not directly
associated with the main form's object. This is achieved by
introducing a `detached?` method to indicate whether an element
is detached and modifying the `html_value` method to retrieve
the value from the input options for detached elements.

Adds a delete element to subforms that are persisted by adding
an `add_delete_element` method that includes a hidden input.
This element represents a "destroy" flag for persisted objects.